requirements.txt
boto3>=1.34.0
datasets>=2.14.0
torch>=2.0.0
unsloth>=2024.1.0
trl>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
huggingface-hub>=0.20.0
jupyter>=1.0.0
pandas>=2.0.0
pyarrow>=14.0.0
requests>=2.31.0
python-dotenv>=1.0.0